requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework
django-cors-headers
django-filter
Markdown
whitenoise